requirements (5).txt
langchain
streamlit
langchain-community
langchain-text-splitters
python-dotenv
faiss-cpu
langchain-groq
PyPDF2
sentence-transformers
pypdf
requests
pdfplumber
langchain-core
langchain-openai
pydantic
unstructured[all-docs]
pillow
lxml
poppler-utils
langchain_astradb
langgraph
duckduckgo-search
pyowm
arxiv
yfinance
xmltodict